infrastructure 層のエラーの定義を domain 層に移動
