+++ infrastructure/lgtm_image_repository.go
@@ -2,7 +2,6 @@ package infrastructure
 
 import (
 	"context"
-	"fmt"
 	"time"
 
 	db "github.com/nekochans/lgtm-cat-api/db/sqlc"
@@ -19,22 +18,13 @@ func NewLgtmImageRepository(db *db.Queries) *lgtmImageRepository {
 	return &lgtmImageRepository{db: db}
 }
 
-type LgtmImageError struct {
-	Op  string
-	Err error
-}
-
-func (e *LgtmImageError) Error() string {
-	return fmt.Sprintf("lgtmImageRepository: %s, %s", e.Op, e.Err)
-}
-
 func (r *lgtmImageRepository) FindAllIds(c context.Context) ([]int32, error) {
 	ctx, cancel := context.WithTimeout(c, dbTimeoutSecond*time.Second)
 	defer cancel()
 
 	ids, err := r.db.ListLgtmImageIds(ctx)
 	if err != nil {
-		return nil, &LgtmImageError{
+		return nil, &domain.LgtmImageError{
 			Op:  "FindAllIds",
 			Err: err,
 		}
@@ -61,7 +51,7 @@ func (r *lgtmImageRepository) FindByIds(c context.Context, ids []int32) ([]domai
 
 	rows, err := r.db.ListLgtmImages(ctx, listLgtmImagesParams)
 	if err != nil {
-		return nil, &LgtmImageError{
+		return nil, &domain.LgtmImageError{
 			Op:  "FindByIds",
 			Err: err,
 		}
